+++ packages/react/components/button/package.json
@@ -43,6 +43,7 @@
   "devDependencies": {
     "@project44-manifest/react-layout": "^1.0.1",
     "@project44-manifest/react-types": "^0.2.1",
+    "@react-types/shared": "^3.16.0",
     "react": "^18.1.0",
     "react-dom": "^18.1.0"
   },
