+++ packages/react/components/button/src/Button/index.ts
@@ -1,0 +1,2 @@
+export * from './Button';
+export * from './Button.types';
