+++ packages/react/components/button/src/ButtonGroup/ButtonGroup.context.ts
@@ -1,5 +1,5 @@
 import { createContext } from '@project44-manifest/react-utils';
-import type { ButtonSize, ButtonVariant } from './Button.types';
+import type { ButtonSize, ButtonVariant } from '../types';
 
 export interface ButtonGroupContext {
   isDisabled?: boolean;
