+++ packages/react/components/button/src/IconButton/index.ts
@@ -1,0 +1,2 @@
+export * from './IconButton';
+export * from './IconButton.types';
