+++ packages/react/components/button/src/index.ts
@@ -1,6 +1,6 @@
 export * from './Button';
-export * from './Button.types';
 export * from './ButtonGroup';
-export * from './ButtonGroup.types';
 export * from './IconButton';
-export * from './IconButton.types';
+export { useButton } from '@react-aria/button';
+export { FocusRing, FocusScope, useFocusRing } from '@react-aria/focus';
+export { useFocus, useHover } from '@react-aria/interactions';
