+++ packages/react/components/button/src/types.ts
@@ -1,0 +1,3 @@
+export type ButtonSize = 'medium' | 'small';
+
+export type ButtonVariant = 'brand' | 'danger' | 'primary' | 'secondary' | 'tertiary';
